requirements.txt
django
psycopg2-binary
pillow
djangorestframework
python-decouple
dj-rest-auth
drf-yasg
django-filter